+++ app/internal/profile/repository.go
@@ -21,13 +21,13 @@ func (r *Repository) Create(ctx context.Context, p *Profile) error {
 	return err
 }
 
-func (r *Repository) FindByID(ctx context.Context, id int64) (*Profile, error) {
+func (r *Repository) FindByID(ctx context.Context, id int) (*Profile, error) {
 	p := new(Profile)
 	err := r.db.NewSelect().Model(p).Where("p.profile_id = ?", id).Scan(ctx)
 	return p, err
 }
 
-func (r *Repository) FindByStoreAndPin(ctx context.Context, storeID int64, pin string) (*Profile, error) {
+func (r *Repository) FindByStoreAndPin(ctx context.Context, storeID int, pin string) (*Profile, error) {
 	p := new(Profile)
 	err := r.db.NewSelect().Model(p).
 		Where("p.store_id = ?", storeID).
@@ -37,17 +37,17 @@ func (r *Repository) FindByStoreAndPin(ctx context.Context, storeID int64, pin s
 	return p, err
 }
 
-func (r *Repository) Delete(ctx context.Context, id int64) error {
+func (r *Repository) Delete(ctx context.Context, id int) error {
 	_, err := r.db.NewDelete().Model(&Profile{}).Where("profile_id = ?", id).Exec(ctx)
 	return err
 }
 
-func (r *Repository) DeleteByStoreID(ctx context.Context, storeID int64) error {
+func (r *Repository) DeleteByStoreID(ctx context.Context, storeID int) error {
 	_, err := r.db.NewDelete().Model(&Profile{}).Where("store_id = ?", storeID).Exec(ctx)
 	return err
 }
 
-func (r *Repository) PinExistsInStore(ctx context.Context, storeID int64, pin string) (bool, error) {
+func (r *Repository) PinExistsInStore(ctx context.Context, storeID int, pin string) (bool, error) {
 	exists, err := r.db.NewSelect().Model(&Profile{}).
 		Where("store_id = ?", storeID).
 		Where("pin = ?", pin).
